requirements.txt
openai~=1.101.0
pandas~=2.3.2
sentence-transformers
numpy~=2.2.6
scikit-learn~=1.7.1
chromadb~=1.0.20
hdbscan~=0.8.40
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
openai>=1.40.0
pyarrow
streamlit
requests
